Reorganized info in tuple.ex
While this `@moduledoc` has the proper information, it was organized in a manner conducive to TLDR. I've tried to put the main tripping points for beginners at the top.

Honestly, the remaining content is somewhat redundant. Implementation details are good to include here but an explanation of a tuple and its most basic functionality is best reserved for the elixir-lang.org documentation. It does not add value for beginners and the functions' documentation are sufficient to explain what they do.