First stab at implementation of `after_suite/1` callback
This is a first take at the `after_suite/1` callback in ExUnit. I've
done basic testing by making this work with an app on my local machine,
but I'm going to think about automated tests tomorrow. I also still
need to document this better, but I wanted to get this up just to make
sure this implementation is along the lines of what y'all are looking
for.