Ensure that mix --cover displays correct coverage in an umbrella app (#8176)
`:cover.start/0` initializes the internal database where coverage data
on all compiled modules is stored. Prior to this change, the database
wasn't cleared out between sub-apps in an umbrella app. When
`:cover.analyse/2` is called, it runs against the entire contents of the
internal database. In an umbrella app, the coverage report for App1
would show App1, the report for App2 would show App1 and App2, App3
would have App1, App2, and App3, and so on.

Closes #8086