Fix flaky test in "mix compile.elixir"
This is an attempt to fix the flaky test:
https://travis-ci.org/elixir-lang/elixir/jobs/437950338

The state of the `:elixir_code_server` is not cleaned up before
the next test is run. The parallel compiler kills all the workers
in case of a compilation error, and the `:elixir_code_server`
cleans the state accordingly ([because it's monitoring those
workers][1]). However, due to race conditions, the next test might
be executed before the `:DOWN` message arrives.

To avoid this scenario, the file `lib/b.ex` is removed since it's
not used in the test "returns error diagnostics".

[1]: https://github.com/elixir-lang/elixir/blob/master/lib/elixir/src/elixir_code_server.erl#L130-L131