Check for line ending in output messages (#8319)
Check for line ending in output messages in ex_unit_tests.exs and beginning of line when applicable. This makes sure we check for the full line,
avoiding unexpected results appearing at the end of the line. Related discussion:
https://github.com/elixir-lang/elixir/pull/8312#discussion_r226867015

Remove on_exit_reload_config/0 and replace it with configure_and_reload_on_exit/1