Allow posix time on touch and document proper defaults
This fixes issue #8450 which was caused by the user of
`system_time` when `os_time` should have been used.

This commit improves the touch API by also allowing
posix times in `File.touch/2` and also document its
defaults. With this change, we phase out the remaining
use of Erlang's `:calendar` module in Elixir.