No longer redirect cross node Logger messages
Prior to this patch, any command initiated by a remote
node would no be logger. So, for instance, if a node
executed a remote command such as "MyApp.prepare_for_shutdown"
and the function above executed some code and logged,
the logged would not appear in the node shutting down,
but instead it would be fully redirect to the caller,
which could be logging or not at all.

This makes sure the message always stays on the current
node. If you need output redirecting, then using I/O is
the best mechanism.