Add ability to filter by more than one line number
As [discussed in the mailing list](https://groups.google.com/d/msg/elixir-lang-core/7yejSvr0RMM/Cfv5JXDnBgAJ), here's a PR that will allow a
shorthand for filtering ExUnit tests by more than one line number at a
time.

For now this doesn't address the question of running single tests in
more than one file since that would be a much larger change. If we do
want to allow that, we can add it later on, but it should be an addition
to the work here, not a replacement.