Extract types from GenServer callbacks
The possible returns for the asynchronous callbacks in a GenServer
(`handle_cast`, `handle_info` and `handle_continue`) were all the same,
but that wasn't easy to see as they all had their own type definition
for the return value in the callback definition.

Also, this will make adding `@spec` annotations to modules implemeting
the `GenServer` behaviour much easier, as they'll now have public types
in `GenServer` that can be used in the same way that
`GenServer.on_start` is often used now.