Support plain regexes. Make it possible to disable regex precompilation.
By default regex strings are compiled on code compilation stage.
This makes it harder to produce packages which would work with different
erlang versions and on different platforms.
Also the runtime prints a warning message when starting if code is running
on a system with different endianness. This can make users anxious.

This can be solved on the code level by recompiling regexes in runtime, but
that requires code hygiene, which may be harder to maintain and also does not
protect from regexes in dependencies.

This pull requests proposes to have a project build configuration, which will
disable regex precompilation by having a plain binary in the regex structure.

Binary works well with `re` functions, all except `inspect`, so there is not
that much changes required.

I'm not sure about specific way to pass the configuration or the name of this
parameter.

Also I wonder if there is a way to mark the code package containing precompiled
regexes to print or suppress the endiannes warning.