Improve performance of release tests (#9112)
This test copied almost the entire Erlang installation in order
to create the scenario for the test. This process was really
slow, to the point that Windows builds in the CI were failing
because the copy was taking longer than 60 seconds to 
complete.

Fortunately, the test do not require all the files. This commit
changes the test to strictly copy the files required for the test.