Add optional inspect/2 callback for Calendar Date, DateTime and NaiveDateTime implement the `Inspect` protocol with a specific implementation for `Calendar.ISO`.
This commit introduces an optional callback, `inspect/2`
that allows calendar implementers to define an
implementation of inspect.

The protocol implementations in `Date`, `DateTime` and
`NaiveDateTime` will invoke these implementations if:

1. The calendar module is loaded
2. The `inspect/2` function is defined in the calendar module

Otherwise the `Inspect.Any.inspect/2` function is
called.

Calendars implementing `inspect/2` should ensure that
the implementation caters for `Date`, `DateTime` and
`NaiveDateTime` structs.