Remove extraneous guard from String.split
This clause makes sense for `pattern` satisfying any of the types
defined in the spec, that is `binary`, `[binary]` and `:binary.cp()`.
Furthermore the guard for `pattern` is not specified in any of the other
clauses.

Additionally, the `or` in the guard seemed a bit out of place, as this
clause (and all others) needs the `string` to be binary to work.

So let's keep just the guard to check the type of the `string` and let
the `:binary.split/3' handle the types of the `pattern`.