Add Keyword.pop!/2 and Map.pop!/2
These function behave like their non-bang counterparts except that
they raise when the key is not present. This can be useful for
option handling in order to extract an option and raise at the same
time if it's not there. Note that this is mostly useful when not
having the key there is an unrecoverable error (for example,
because you already validated options): if you want to return a
nice error to the user, "pop_lazy/3" with a "raise" in the
third-argument function might be a better idea.