Limit which directories from lib are included
Some of the directories inside of lib may be from previous builds. This commit
walks the included apps and ensures that only they are include in the build. The
test explicitly creates an old application and ensures that it is not included
in the build.

All of the files (not directories) inside of the releases dir are also included
in the tar, the specific directory for the version of erts is explicitly added.