Ensure project's Logger config is used when running Mix tasks
Part of the configuration of the Logger can be changed
dynamically but other part is only used at start. For example,
the `:console` backend is implemented with `gen_event`, and
its configuration is only read on `init` to build the state.

Logger is first started by Elixir using its default
configuration, but the project's configuration is not used
fully until the Logger application is restarted. This usually
happens when the application is started, but not all Mix tasks
do such thing.

This commit ensures that Logger is restarted before calling any
Mix task.