Add compare support to Enum.sort and add Enum.sortt_reverse*
Before this PR, sorting dates, decimals, and other structs was
somewhat awkward:

    Enum.sort(dates, &(Date.compare(&1, &2) != :lt)) # ascending
    Enum.sort(dates, &(Date.compare(&1, &2) != :gt)) # descending

Sorting a user by dates, decimals, and other structs was equally
verbose:

    Enum.sort_by(users, &(&1.birthday), &(Date.compare(&1, &2) != :lt)) # ascending
    Enum.sort_by(users, &(&1.birthday), &(Date.compare(&1, &2) != :gt)) # descending

Even a general sort in reverse required passing &>=/2 as an argument,
which is not clear in intent:

    Enum.sort(list, &>=/2) # descending

This PR addresses this by adding two changes:

  1. Sorting funtions now accept a module as sorter. The compare/2
     function of said modules are then used for sorting.

  2. `sort_reverse*` functions were added to complement the
     functionality above

All of the examples above can be written as:

    Enum.sort(dates, Date) # ascending
    Enum.sort_reverse(dates, Date) # descending
    Enum.sort_by(users, &(&1.birthday), Date) # ascending
    Enum.sort_reverse_by(users, &(&1.birthday), Date) # descending
    Enum.sort_reverse(list) # descending

The reason why we picked `sort_reverse` instead of `sort_descending`
because a sorting funtion can still be given to `sort_reverse`, which
means a descending function will return an ascending result (i.e. the
reverse).