Diff strings with escaping with empty strings as well
String.bag_distance/2 always returns 0.0 if one of the strings is empty,
but that's probably not the behavior that we want in ExUnit, so I
enabled full diffing on strings if one of the two is empty.

This fixes the reported bug, but I'm not sure it's the right way to fix
this, so if folks have any other input I'd love to hear it.