Correct title of section
In 6fb8efc6482d5c8526c507b91cd0982d221ec08c I accidentally
changed the title of the section
"OS Processes Partitioning" to
"Operating system process partitioning"
(notice "process" is in singular). Since "partitioning"
is the noun, and the rest are adjective, I think "process"
should be in singular.

This PR changes the reference to the section with the new title.