use expression idiomatically
"Over the course of the next section" sounds more natural to me
than "over the next section".  "Over the next _________" usually
has a temporal expression to fill in the blank, like "over the
next 3 days" or "over the summer".

"Throughout the next section" or "In the next section" would also
make sense and flow naturally.