insert definite article
"many of Elixir syntax constructs" could be changed in a few different ways:

* "many ~of~ Elixir syntax constructs"  (remove the "of)
* "many of Elixir's syntax constructs"
* "many of the Elixir syntax constructs"

I felt the first one was clearest in both text and speech.
(Having two adjacent S sounds with _Elixir's syntax_ is harder
to understand if the text is read aloud.)  If you prefer another
and want me to change the PR I'd be happy to.