Nil-safe list traversals in Access helper functions
This is a follow-up to my elixir-lang-core thread with the
subject "PROPOSAL: Nil-safe Access.at/1"

I've broadened beyond just `at/1` to achieve the consistency
discussed in that thread. Changed functions (all in Access)
are: `at/1`, `elem/1`, `all/0`, `filter/1`

Note: I did not modify `key/1,2` as had been suggested in the
thread because that deals with maps and structs and my focus was
initially about List traversals, but I'm happy to add that if
there's interest.

I've also only implemented the changes for the `:get` case and
deferred `:get_and_update` mostly due to familiarity. I rarely
use `Kernel.get_and_update_in` so I particularly welcome more
discussion of the desirability of nil-safety in those cases.