Add application tracers to Mix
This tracer is used to validate that all modules being invoked
in an application belongs to an application that is currently
listed as a dependency. This will help enforce umbrella application
boundaries and help users discover when they depend on Elixir and
Erlang libraries without declaring them upfront.

Thanks to @sasajuric for the proof of the concept and insights
that made this feature possible.

Closes #9847.