Load vm-args for remote shell
We're writing our own empd/distribution module, similar to https://www.erlang-solutions.com/blog/erlang-and-elixir-distribution-without-epmd.html .

In order to be able to attach to a remote console, we need to load the same `rel/vm.args` file as the running instance of the beam. I'm not sure if this would cause any unintended side effects.