Add column information to assigns in EEx templates
Before this patch we had:

    iex> Code.put_compiler_option(:parser_options, columns: true)
    iex> {_, _, [{_, _, [_, {_, _, [x]}]}, _]} = EEx.compile_string("<%= foo(@hi) %>"); x
    {:foo, [line: 1, column: 5],
     [
       {{:., [line: 1],
         [{:__aliases__, [line: 1, alias: false], [:EEx, :Engine]}, :fetch_assign!]},
        [line: 1],
        [
          {:var!, [line: 1, context: EEx.Engine, import: Kernel],
           [{:assigns, [line: 1], EEx.Engine}]},
          :hi
        ]}
     ]}

Note, no column information on the assign access.

After this patch we have:

    {:foo, [line: 1, column: 5],
     [
       {{:., [line: 1],
         [{:__aliases__, [line: 1, alias: false], [:EEx, :Engine]}, :fetch_assign!]},
        [column: 9, line: 1],
        [
          {:var!, [line: 1, context: EEx.Engine, import: Kernel],
           [{:assigns, [line: 1], EEx.Engine}]},
          :hi
        ]}
     ]}

(note: `[line: ..., column: ...]` vs `[column: ..., line: ...]` hence
sorting in the test.)

I'm not sure if other nodes in the AST should also get the column,
I suppose it's fine to add it to the `EEx.Engine` one, but not sure if
it is appropriate to add it to the `fetch_assign!` and `var!` ones
as the column offset would be artificial. For the reason, maybe we
shouldn't change the column information after all.