Add app.load task
Prior to this commit, application loading was scattered
around Mix, which lead to inconsistent behaviour.

We now always load the applications before compiling,
simplifying the logic on both app.start and in the
application tracer. This also paves the road to
config/runtime.exs, which needs all apps loaded
before they are started.