Add MIX_BUILD_ROOT to config _build dir
The goal is to make it easier to change the path of the "_build"
directory.

Mix currently supports MIX_BUILD_PATH to customize where the build
artifacts are written to. One thing to keep in mind is that
MIX_BUILD_PATH expects the final path and it's up to the user to
use different paths per environment. MIX_BUILD_PATH is not
equivalent to the "_build" directory, it's equivalent to the
"_build/<env>" directory.

Mix also allows to configure the path of the "_build" directory
using the :build_path configuration of Mix.Project, but that is
meant to be used by umbrella applications:

    In a non-umbrella context, configuring this has undesirable
    side-effects (such as skipping some compiler checks) and
    should be avoided.

Finally, Mix.Project also uses :env_path to configure the build
path, but that option is not documented and it's marked as
"private" in the code.