Introduce a type hint system
Instead of trying to assuming that a certain usage of
a type is the correct one, we keep the unification message
agnostic and add hints to the message.

Each message may have multiple hints, as they are computed
alongside the trace.

Here is how the new warning for dot access look like:

    warning: incompatible types:

        map() !~ atom()

    in expression:

        # a.ex:4
        module.__struct__

    where "module" was given the type map() (due to calling var.field) in:

        # a.ex:4
        module.__struct__

    where "module" was given the type atom() in:

        # a.ex:3
        %module{}

    HINT: "var.field" (without parentheses) implies "var" is a map() while
    "var.fun()" (with parentheses) implies "var" is an atom()

    Conflict found at
      a.ex:4: A.a/1