Improve docs and argument names for Collectable module
It is not clear what original mean, if the element we extract data or the element we put into.
For clarity I chose to name it the same as the module.

this change make it clear that what needs to be returned is the initial accumulator and not the collectable we are
puting data into.

-          {original, collector_fun}
+          initial_acc = map_set
+
+          {initial_acc, collector_fun}

mention that the MapSet implementation is a simplified version, not the real one.

The implementations in the module have been reformated and some attribute names changed, so it is easier for developers who want to read the source code to understand how protocols are implemented.