Update autocomplete.ex
With this change when we fail to find docs for erlang functions we will now try to reconstruct the signature from its own specs. Note that in the "case do" inside format_definition_from_typespec when no match is found we just pass the whole spec (not sure if we want to keep this in the final version).