Add :turkic mode option to String case functions
This PR adds support of Turkic languages casing of the letter i according to Unicode SpecialCasing.txt.

It left properly the standard handling of the letter i with a dot above.

Fix according to Unicode SpecialCasing.txt

Revert back to original the Unicode SpecialCasing.txt
Add support to the combining dot character
Changes the option name to :turkic