Do not call Logger.configure on config_change callback.
The `config_change` callback is called from `application_controller` process,
which is receiving env updates on `Application.put_env`.
`Logger.configure` does a blocking call to the `Logger` process, which calls
`Application.put_env`

This causes a crash on logger config update.
If logs are synchronous, failing logger will call itself to report a crash and get stuck.

Main purpose of Logger.configure is to put env into the app config and
`config_change` is called when env is already updated.

To properly handle `config_change` if that's necessary,
the app should call some other endpoint which will not call `Application.put_env`