Move printing test seed to beginning of test run
Came across this proposal on the forums to move the seed of the
test suite to the beginning of the test run. This was actually
something I wished for recently when encountering a weird issue
in our codebase which caused all of our non async tests to fail and
each synchronous test would take 60 seconds before timing out. With
a few hundred synchronous tests this took quite some time for the
suite to finish and I wanted to grab the seed to replicate it.

https://groups.google.com/g/elixir-lang-core/c/sP8DiY1zMFQ